Refactor Palette to core/palette/
