+++ foundry/game/level/WorldMap.py
@@ -1,10 +1,10 @@
 from PySide6.QtCore import QPoint, QSize
 
+from foundry.core.palette.Palette import PaletteGroup
 from foundry.game.File import ROM
 from foundry.game.gfx.drawable.Block import Block
 from foundry.game.gfx.GraphicsSet import GraphicsSet
 from foundry.game.gfx.objects.MapObject import MapObject
-from foundry.game.gfx.Palette import PaletteGroup
 from foundry.game.level.LevelLike import LevelLike
 from foundry.smb3parse.levels.world_map import (
     WORLD_MAP_HEIGHT,
