+++ foundry/gui/BlockViewer.py
@@ -11,12 +11,12 @@ from PySide6.QtGui import (
 from PySide6.QtWidgets import QComboBox, QLabel, QLayout, QStatusBar, QToolBar, QWidget
 
 from foundry import icon
+from foundry.core.palette.Palette import PALETTE_GROUPS_PER_OBJECT_SET, PaletteGroup
 from foundry.core.Position import Position
 from foundry.core.UndoController import UndoController
 from foundry.game.File import ROM
 from foundry.game.gfx.drawable.Block import Block
 from foundry.game.gfx.GraphicsSet import GraphicsSet
-from foundry.game.gfx.Palette import PALETTE_GROUPS_PER_OBJECT_SET, PaletteGroup
 from foundry.gui.BlockEditor import BlockEditorController as BlockEditor
 from foundry.gui.CustomChildWindow import CustomChildWindow
 from foundry.gui.LevelSelector import OBJECT_SET_ITEMS
@@ -95,7 +95,7 @@ class BlockViewerController(CustomChildWindow):
         self.palette_group_changed.connect(self.palette_group_spinner.setValue)
 
         self.toolbar.addWidget(self.bank_dropdown)
-        self.toolbar.addWidget(QLabel(" Object Palette: "))
+        self.toolbar.addWidget(QLabel(" Object palette: "))
         self.toolbar.addWidget(self.palette_group_spinner)
 
         self.addToolBar(self.toolbar)
